Bug fix for debug builds in micro_session.cc (#6968)
* If the build decides not to inline kReceiveBufferSizeBytes,
  we will encounter a linking error.

+++ src/runtime/micro/micro_session.cc
@@ -148,9 +148,10 @@ class MicroTransportChannel : public RPCChannel {
             ::std::max(::std::chrono::microseconds{0},
                        ::std::chrono::duration_cast<::std::chrono::microseconds>(
                            end_time - ::std::chrono::steady_clock::now()))};
-        chunk = frecv_(kReceiveBufferSizeBytes, iter_timeout.count()).operator std::string();
+        chunk =
+            frecv_(size_t(kReceiveBufferSizeBytes), iter_timeout.count()).operator std::string();
       } else {
-        chunk = frecv_(kReceiveBufferSizeBytes, nullptr).operator std::string();
+        chunk = frecv_(size_t(kReceiveBufferSizeBytes), nullptr).operator std::string();
       }
       pending_chunk_ = chunk;
       if (pending_chunk_.size() == 0) {
